Update `url` in Credentials Object Example to match spec
The description for the `url` field in a `Credentials` object reads _“The URL to your API versions endpoint.”_
The contained example following it shows `https://example.com/ocpi/cpo/` as a value for the `url` field.

When checking the documentation on the Versions Endpoint, it is always mounted on `https://…/versions`. This change updates the value of `url` to be in line to that part of the documentation.