Update tariff_14_step_size.json
"24:00" is not an acceptable time value.  Databases will convert it to "00:00", which is not what's intended.
This does bring up the question of how to we describe "midnight the next day" as a time value that we can use?
Also, what happens if someone happens to cause an event at 23:59:30?  Hopefully that either gets cut off to be "23:59" or rounds up to "00:00" of the next day.