Added use case for failed authorization
This is a specific use case where driver needs to act in order to restore validity of the token. For example, there's no prepaid credit left. We should tell the user what is wrong in a language they understand. This requires a charge point with a display or a push message from the MSP app. 

The use case could be numbered in a way that indicates it's not a core requirement. (I'm working on this topic currently in my implementation of OCPI 2.2 and it's quite tricky)