Fix typo unqiue -> unique
A couple of diagrams in the chapter 4, subsection 2, 4.2. Unique message IDs
had a type in the name of the diagram, was not a problem because was
include in the document with the type.

This fix, do not bring any "Value" to the document, but is just something
which can be fix, to have a good document.